refactor: Move factories into a
